megaco_codec_meas (megaco v4.6)

This module implements a simple megaco codec measurement tool.

Results are written to file (excel compatible text files) and on stdout.

Note that this module is not included in the runtime part of the application.

Functions

start()

Equivalent to start/1

start/1

start(MessagePackage) -> void()

This function runs the measurement on all the official codecs; pretty, compact, ber, per and erlang.