megaco_tcp (megaco v4.6)

Interface module to TPKT transport protocol for Megaco/H.248.

This module contains the public interface to the TPKT (TCP/IP) version transport protocol for Megaco/H.248.

Functions

block(Socket)

block(Handle) -> ok

Stop receiving incoming messages on the socket.

close(Socket)

close(Handle) -> ok

This function is used for closing an active TPKT connection.

connect(SupPid, Parameters)

connect(TransportRef, OptionList) -> {ok, Handle, ControlPid} | {error, Reason}

This function is used to open a TPKT connection.

module - This option makes it possible for the user to provide their own callback module. The receive_message/4 or process_received_message/4 functions of this module is called when a new message is received. Which one is called depends on the size of the message;

small - receive_message

large - process_received_message

Default value is megaco.

inet_backend - Choose the inet-backend.

This option make it possible to use a different inet-backend ('default', 'inet' or 'socket').

Default is default (system default).

get_stats()

Equivalent to get_stats/2

get_stats(Socket)

Equivalent to get_stats/2

get_stats(Socket, Counter)

get_stats(SendHandle, Counter) -> {ok, CounterStats} | {error, Reason}

Retreive the TCP related (SNMP) statistics counters.

listen(SupPid, Parameters)

listen(TransportRef, ListenPortSpecList) -> ok

This function is used for starting new TPKT listening socket for TCP/IP. The option list contains the socket definitions.

inet_backend - Choose the inet-backend.

This option make it possible to use a different inet-backend ('default', 'inet' or 'socket').

Default is default (system default).

reset_stats()

Equivalent to reset_stats/1

reset_stats(Socket)

reset_stats(SendHandle) -> void()

Reset all TCP related (SNMP) statistics counters.

send_message(Socket, Data)

send_message(Handle, Message) -> ok

Sends a message on a connection.

socket(Socket)

socket(Handle) -> Socket

This function is used to convert a socket_handle() to a inet_socket(). inet_socket() is a plain socket, see the inet module for more info.

start_transport()

start_transport() -> {ok, TransportRef}

This function is used for starting the TCP/IP transport service. Use exit(TransportRef, Reason) to stop the transport service.

unblock(Socket)

unblock(Handle) -> ok

Starting to receive incoming messages from the socket again.

upgrade_receive_handle(Pid, NewHandle)

upgrade_receive_handle(ControlPid, NewHandle) -> ok

Update the receive handle of the control process (e.g. after having changed protocol version).