megaco_udp (megaco v4.6)

Interface module to UDP transport protocol for Megaco/H.248.

This module contains the public interface to the UDP/IP version transport protocol for Megaco/H.248.

Functions

block/1

block(Handle) -> ok

Stop receiving incoming messages on the socket.

close/1

close(Handle) -> ok

This function is used for closing an active UDP socket.

create_send_handle/3

create_send_handle(Handle, Host, Port) -> send_handle()

Creates a send handle from a transport handle. The send handle is intended to be used by megaco_udp:send_message/2.

get_stats()

Equivalent to get_stats/2

get_stats(SH)

Equivalent to get_stats/2

get_stats(SH, Counter)

get_stats(SendHandle, Counter) -> {ok, CounterStats} | {error, Reason}

Retreive the UDP related (SNMP) statistics counters.

open(SupPid, Options)

open(TransportRef, OptionList) -> {ok, Handle, ControlPid} | {error, Reason}

This function is used to open an UDP/IP socket.

module - The option makes it possible for the user to provide their own callback module. The functions receive_message/4 or process_received_message/4 of this module is called when a new message is received. Which one depends on the size of the message:

small - receive_message

large - process_received_message

Default value is megaco.

inet_backend - Choose the inet-backend.

This option make it possible to use a different inet-backend ('default', 'inet' or 'socket').

Default is default (system default).

reset_stats()

Equivalent to reset_stats/1

reset_stats(SH)

reset_stats(SendHandle) -> void()

Reset all TCP related (SNMP) statistics counters.

send_message/2

send_message(SendHandle, Msg) -> ok

Sends a message on a socket. The send handle is obtained by megacoudp:create_send_handle/3. Increments the NumOutMessages and NumOutOctets counters if message successfully sent. In case of a failure to send, the NumErrors counter is _not incremented. This is done elsewhere in the megaco app.

socket/1

socket(Handle) -> Socket

This function is used to convert a socket_handle() to a inet_socket(). inet_socket() is a plain socket, see the inet module for more info.

start_transport()

start_transport() -> {ok, TransportRef}

This function is used for starting the UDP/IP transport service. Use exit(TransportRef, Reason) to stop the transport service.

unblock/1

unblock(Handle) -> ok

Starting to receive incoming messages from the socket again.

upgrade_receive_handle(Pid, NewHandle)

upgrade_receive_handle(ControlPid, NewHandle) -> ok

Update the receive handle of the control process (e.g. after having changed protocol version).